services/file-parser-service/requirements.txt
flask
flask-cors
python-dotenv
PyPDF2
python-pptx
